aos/ability/product_control/requirements.txt
tinydb
pyzmq
grpcio
grpcio-tools
Flask
gunicorn